refactor: 불필요한 함수 삭제
